drivers/sx126x: fix NETOPT_RX_SYMBOL_TIMEOUT

+++ drivers/include/sx126x.h
@@ -107,7 +107,7 @@ struct sx126x {
     sx126x_pkt_params_lora_t pkt_params;    /**< Lora packet parameters */
     sx126x_mod_params_lora_t mod_params;    /**< Lora modulation parameters */
     uint32_t channel;                       /**< Current channel frequency (in Hz) */
-    uint8_t rx_timeout;                     /**< Rx Timeout in terms of symbols */
+    uint16_t rx_timeout;                    /**< Rx Timeout in terms of symbols */
     bool radio_sleep;                       /**< Radio sleep status */
 };
